C++ compatibility for the code a wider configuration enables.
Same treatment as the earlier clang++ pass, extended to the sources that
were compiled but dead under the old configuration: floats, mpz integers,
sets, slices, unicode strings, property, enumerate/filter/reversed,
min/max, detailed error reporting, pystack and computed goto dispatch.

All of it is mechanical and behaviour preserving:
  - explicit casts on the void* results of MP_OBJ_TO_PTR() and
    mp_pystack_alloc(), which C converts implicitly and C++ does not
  - explicit casts where an int is passed as an mp_unary_op_t or
    mp_binary_op_t, and on two union initializers binary.c narrows from
    long long
  - two goto fixes: objint_mpz.c declares "res" without an initializer so
    the zero-division goto may enter its scope, and objtype.c gains a brace
    block so the skip_special_accessors goto jumps past the property lookup
    declarations rather than into them

+++ py/binary.c
@@ -214,10 +214,10 @@ mp_obj_t mp_binary_get_val(char struct_type, char val_type, byte **ptr) {
         return mp_obj_new_str(s_val, strlen(s_val));
 #if MICROPY_PY_BUILTINS_FLOAT
     } else if (val_type == 'f') {
-        union { uint32_t i; float f; } fpu = {val};
+        union { uint32_t i; float f; } fpu = {(uint32_t)val};
         return mp_obj_new_float(fpu.f);
     } else if (val_type == 'd') {
-        union { uint64_t i; double f; } fpu = {val};
+        union { uint64_t i; double f; } fpu = {(uint64_t)val};
         return mp_obj_new_float(fpu.f);
 #endif
     } else if (is_signed(val_type)) {
